Moved menus to a separate class to better manage them. **May need to create individual classes and inherit them from menu**
